requirements.txt
mysql-connector-python
# Pillow
flask

# Data Analysis and Reporting
# numpy
# pandas
# matplotlib
# fpdf
